refactor: adopt Protocol-based Source/Sink with strict mypy
- EventSource/EventSink: ABC → Protocol (structural typing)
- EventSink.send: dict → Event
- Event.timestamp: float → int (nanoseconds since boot)
- ReplayEventSource: yield raw str → parse NDJSON, yield Event
- EBPFEventSource: add imports, yield typed Event
- KafkaEventSink: add imports, serialize via asdict
- Tests: remove _parse helper, add full type annotations
- Add mypy>=1.10 dev dep, [tool.mypy] strict=true
- Add CLAUDE.md with uv commands and conventions

+++ event.py
@@ -1,8 +1,9 @@
 from dataclasses import dataclass
 
+
 @dataclass(frozen=True)
 class Event:
-    timestamp: float
+    timestamp: int  # nanoseconds since boot
     pid: int
     process: str
     payload: str
